+++ src/onegov.core/onegov/core/__init__.py
@@ -7,7 +7,7 @@ from onegov.core.framework import Framework
 # include the directive module so it gets picked up by the morepath config
 from onegov.core import directive  # noqa
 
-# include the jsx module so it gets picked up by webassets
-from onegov.core import jsx  # noqa
+# include the filters module so they get picked up by webassets
+from onegov.core import filters  # noqa
 
 __all__ = ['Framework', 'log']